Use C-type (not index terms) for step in qpred
The previous commit introduce an error for a using array_shift on a void
pointer, but the location information was not useful. This was because
the array_shift was being deconstructed into a base pointer, and a
sizeof term, and the latter term, by virtue of being generated by CN
developers than the user, had an 'other' location. When this term was
later checked in wellTyped, it would error.

Feeling uncomfortable about lying about the origin and made-up nature of
the sizeof term, I instead did a long-overdue refactor of QPredicate.t
to have its "step" field just be a C-type instead of an arbitrary index
term.

I believe enough checks and errors about this have been in place about
this for long enough such that this is non-breaking change (and I also
ran the tests).