Fix rems-project/cerberus#427, crash on debug >= 5
Cerb_frontend.Pp_symbol.to_string_pretty_cn was updated to (a)
not depend at all on the Cerberus debug level to decide whether or not
to print symbols with an "_{<num>}" suffix, and 9b) take an optional
named argument to control that instead.

This commit updates the CN usage of that API, and fixes a solver crash
in doing so. This is because the SMTLib2 simple-symbol token does not
accept braces as part of symbol names (it does as part of names quoted
with a |_| but that would have made the SMT output more noisy and harder
to read or the code more complex to check whether to escape or not).
It does so by using a separate API which never prints the
number-in-braces prefix for the uses in the solver, even if the flag for
doing so is set for CN.

It also does the same for auto-generated variables for unpacking
predicates and function calls, created in check.ml/typing.ml.

I also suspect a similar bug may affect the Coq output, and so that too
will be changed.