Use dev profile to build locally
For dune, the `-p` and `install` commands (anything related to opam
files) by default use "release" profile, which will have the warnings
enabled but not erroring. This means we were always building with the
release profile, and a bunch of warnings about our code were being
ignored.

This commit changes the Makefile to use the `dev` profile by default,
and fixes up a bunch of warnings (mostly by silencing them since I don't
know the correct fix).

However, the CI still uses the opam file (which uses `-p`, which uses
the release/warnings are just warnings, not errors) and so it will not
break if they get through. On one hand, the CI should emulate what users
will do to install CN, and is conventional within Dune/OCaml. On the
other hand, it's annoying if such warnings get through. So I am unsure
if I should instead change the behaviour of the release profile to
consider warnings as errors.