Add an optional `--cc=` argument common to all commands
This allows the user to specify which external C compiler should be used by
CN. If the argument is absent, the CC env variable is read. If that is not
defined, we fallback to using `cc` (like previously).

This first commit only updates the call to the preprocessor by the
cerberus pipeline.