[CN-exec] Change bump allocator block size and count (#219)
Now that CN maps are allocated using the bump allocator and not the free-list allocator/`malloc`, it was exceeding `BUMP_BLOCK_SIZE` quite easily. 

We keep the maximum space in the bump allocator as 1GB but make the size of each block 8x bigger and the number of blocks 8x smaller.