different CN/solver interaction for SMT counter models
CN used to create a new solver instance for each model. This means solver models
can be passed around as values and interacted with later, but also complicates
the Solver module. With this change, CN models are just references to the
current solver state (so creating a model no longer creates a new solver
instance). This means the timing of when CN queries the model matters, and type
error reports including countermodel information have to be produced when the
model is still "live". This change temporarily disables the CheckPredicates
logic, which relies on asking new solver queries, which then override the
present model -- to be fixed.