Use MemByte in type checking
This commit transitions the to_/from_byte mechanism to using
MemBytes and updates the cn_vip_testsuite along with it.

Note that the representation of MemBytes is not in its final form, and
neither is its use.

The work-around for re-ordering Z3 models by dependency was reinstated,
despite an upgrade to 4.15.2, since one of the tests
(pointer_copy_user_dataflow_direct_bytewise.pass.c) was exhibiting an
the bug (https://github.com/Z3Prover/z3/issues/7268).

Some tests no longer pass because bytes are not treated identically to
unsigned chars by Cerberus. In particular, casting to unsigned chars,
bit-twiddling, and writing literals to them is not supported. This could
be added in the future by extending the Cerberus elaboration and memory
interface with new casts, but for now, it's not a priority.

pointer_from_integer_1p.error.c has changed behaviour independently (I
don't know when) so this commit also updates that.