Add provenance constraints for pointer bytes
This commit adds a first pass at provenance constraints for pointer
bytes. This means that converting a pointer to bytes, and then
converting it back will retain the provenance information it had.

It is a first pass because combine_prov is a bit more complicated than
this but it suffices for now.