Implement combine_prov using SMT constraints
This commit implements combine_prov by splitting it up into two
constraints; first that all the non-CN_None provenances are equal, and
second that a new variable "combined_prov", is equal to the first (in
endian order) non-CN_None provenances. If both these conditions are
satisfied, then the value's provenance must be equal to the latter.
If not, the value's provenance is unconstrained, that is empty.