Change rep of byte provenance to be optional
This commit changes the backend representation of memory bytes to be
Option Alloc_id and then updates the s-expression parsing to handle
that.

Using that, it implements combine_prov by splitting it up into two
constraints; first that all the non-CN_None provenances are equal, and
second that a new variable "combined_prov", is equal to the first (in
endian order) non-CN_None provenances. If both these conditions are
satisfied, then the value's provenance must be equal to the latter.
If not, the value's provenance is unconstrained, that is empty.