Fix non-CN_None bytes check and provenance check
This commit fixes two issues of the previous commit. The first is that
the constraint for all bytes being non-CN_None may be assumed when going
to bytes, but must be checked (with provable) when going From bytes.

The second is that when going From bytes, the check that any non-CN_None
provenances must be equal needs to (sadly) be done all-pairs-wise,
resulting in 36 pairwise constraints (9 * 8 / 2).