Map Byte to Option MemByte
This commit changes the mapping of the Byte C type to Option MemByte
from MemByte and updates all the associated constraints with it.

Note that the constraint for all bytes being non-CN_None may be assumed
when going To bytes, but must be checked (with provable) when going From bytes.

Also, when going From bytes, the check that any non-CN_None
provenances must be equal needs to (sadly) be done all-pairs-wise,
resulting in 36 pairwise constraints (9 * 8 / 2).