Skip label inlining for Fulminate
Milicore label inlining causes an exponential blow-up in some pKVM code
with switches, making it unfeasible to run Fulminate on those examples.

This commit disables Milicore label inlining only for Fulminate to
address this. In doing so, it adds a new type of Mucore label, `Other`,
which contains a dummy specification (argument type) for each label,
which is required to pass the well-formedness checks on Mucore run
before Fulminate.