Skip label inlining for Fulminate  (#319)
Mlicore label inlining causes an exponential blow-up in some pKVM code
with switches, making it unfeasible to run Fulminate on those examples.

This commit disables Milicore label inlining only for Fulminate to
address this. In doing so, it adds a new type of Mucore label, `Non_inlined`,
which contains a dummy specification (argument type) for each label,
which is required to pass the well-formedness checks on Mucore run
before Fulminate.

The switch for deciding to inline or not uses a named
flag, hard-coded by the commands calling the well-formedness check.

---------

Co-authored-by: Rini Banerjee <rb2018@cam.ac.uk>