Use dummy type for undef in non-inlined label run
Not inlining labels means that some label arguments
need to inferred rather than checked. This includes
the UB for reaching the end of a function which returns
a value without having a value to return. In this case,
I just synthesise a Unit type as a dummy if we're in
executable spec mode.