Experimental CLI flag for unions (#382)
* Get wf to accept unions for Fulminate

* Use sized arrays for union work-around

This will allow sizeof(union tag) to work in the pipeline.

* Use dummy type for undef in non-inlined label run

Not inlining labels means that some label arguments
need to inferred rather than checked. This includes
the UB for reaching the end of a function which returns
a value without having a value to return. In this case,
I just synthesise a Unit type as a dummy if we're in
executable spec mode.

* more union workarounds in welltyped.ml

* DM suggestion: map Option<MemByte> to u8

* Transform expr in struct conversion functions if of union type

* add CLI flag and internal global flag

* change test back

* add newline

---------

Co-authored-by: Dhruv Makwana <dc-mak@users.noreply.github.com>
Co-authored-by: Christopher Pulte <cp526@cam.ac.uk>