[CN-exec] Support for GCC statements as expressions (#376)
This PR adds support for GCC-statements-as-expressions, now mapping and unmapping variables declared inside the statements-as-expression block to and from the ownership ghost state. This involves storing the final expression in the GCC-statement-as-expression in a ghost variable so the relevant unmappings can be done, before 'returning' the ghost variable as the value of the entire stat-as-expr.